% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSET.R
\name{COSET}
\alias{COSET}
\title{COnvex subspace Shrinkage via Exponential Tilting}
\usage{
COSET(
  draws0,
  Con_Matrix = diag(ncol(draws0[[1]])),
  lb = -rep(Inf, ncol(draws0[[1]])),
  ub = rep(Inf, ncol(draws0[[1]])),
  CI_level = 0.95,
  model_selection = c("bayes_factor", "waic")[1],
  nu,
  nu_max,
  min_ESS = 1/10 * nrow(draws0$posterior),
  verbose = TRUE,
  cl
)
}
\arguments{
\item{draws0}{named list with names 'posterior', 'prior', and 'logLik' (if using WAIC).
Each ought to be a matrix of posterior, prior, or log likelihood draws (each row is a draw),
all of which should correspond to the base prior.}

\item{Con_Matrix}{Definition of convex subspace which describes prior relationships
between parameters.}

\item{lb}{numeric. Lower bound of convex subspace. Can be \eqn{-\infty}.}

\item{ub}{numeric. Upper bound of convex subspace. Can be \eqn{\infty}.}

\item{CI_level}{numeric in (0,1). Level of the
posterior credible interval.}

\item{model_selection}{character. One of "bayes_factor" or "waic". Method of picking level of
shrinkage via nu.}

\item{nu}{shrinkage parameter.  If missing, the optimal nu via Bayes factors will be used}

\item{nu_max}{maximum value of nu to be considered.}

\item{min_ESS}{integer.  Minimum number of effective sample size in considering values of nu.}

\item{verbose}{logical. Should any output be provided?}

\item{cl}{optional object of class c("SOCKcluster", "cluster") generated from parallel::makeCluster.}
}
\value{
Object of class "coset", with the following named elements:
\itemize{
\item summary, a data frame with the following structure:
\itemize{
\item variable - Variable names extracted from colnames(Sigma0)
\item mean - Posterior mean
\item \link{(1-CI_level)/2}\% - lower CI bound
\item \link{1 - (1-CI_level)/2}\% - upper CI bound
}
\item nu the SUBSET shrinkage parameter either fixed by the user or selected by Bayes factor
\item ESS the effective sample size
\item BF_favoring_nu The Bayes factor in favor of shrinkage vs. no shrinkage (if Bayes Factor was chosen as model selection method)
\item WAIC A vector with the following named elements (if WAIC was chosen as model selection method):
\itemize{
\item WAIC_COSET The WAIC of the model using the COSET prior
\item WAIC_base The WAIC of the model using the base prior
}
\item proposal_draws Draws input by user (posterior draws under the base prior).
\item is_weights Importance sampling weights corresponding to the proposal draws under the
SUBSET prior
\item base_prior_draws Draws input by the user (prior draws under the base prior).
}
}
\description{
COSET() will perform importance sampling
using an exponential tilted prior that shrinks the
parameters towards a convex subspace.
}
